Created a md file for the compiler CFG.
Fixed lexical analysis logic to allow strings and added functionality for higher dimentional arrays.
